Cropping semi working on new structure
Pushing current work, there seems to be a bug with rotations. After I rotated a bunch, the orientation I saved on was slightly different than the one that populated the screen. Not sure if this is an error due to recent changes or legacy from older changes. Will look into more.